remove duplication code smell
