+++ test/delegates/utils/index.spec.ts
@@ -1,7 +1,7 @@
 import cardEntitiesSpec from './card-entities.spec';
+import editorSchema from './editor-schema.spec';
 import deviceSpec from './get-device.spec';
 import hasProblemSpec from './has-problem.spec';
-import integrationSchemaSpec from './integration-schema.spec';
 import isIntegrationSpec from './is-integration.spec';
 
 export default () => {
@@ -10,6 +10,6 @@ export default () => {
     hasProblemSpec();
     deviceSpec();
     isIntegrationSpec();
-    integrationSchemaSpec();
+    editorSchema();
   });
 };
